requirements.txt
Django==1.2.1
django-piston==0.2.2
psycopg2==2.2.2
